requirements.txt
# PyTorch com suporte a CUDA 11.8 (GPU)
torch==2.6.0+cu128
torchvision==0.16.0+cu128
torchaudio==2.1.0+cu128
# Instalar com:
# pip install -r requirements.txt --extra-index-url https://download.pytorch.org/whl/cu118

# FAISS com GPU
faiss-gpu==1.7.2

# Transformers e tokenização
transformers==4.37.2
sentence-transformers==2.2.2
huggingface-hub==0.20.3
tokenizers==0.15.1

# Utilitários
numpy==1.24.4
python-docx==1.1.0
scikit-learn==1.3.2
tqdm==4.66.1

# (opcional, se quiser adicionar interface futuramente)
# streamlit==1.32.2
# gradio==4.20.1


